+++ base/exports.jl
@@ -486,6 +486,7 @@ export
     any,
     firstindex,
     collect,
+    count!,
     count,
     delete!,
     deleteat!,